Fix unsigned long mapped as Java long instead of NativeLong
C unsigned long is platform-dependent (4 bytes on 32-bit, 8 bytes on
64-bit), but Java long is always 8 bytes. This mismatch causes
incorrect struct layout and calling convention issues on 32-bit
systems.

- virDomainBlockJobInfo.bandwidth: long -> NativeLong
  (fixes struct layout so cur/end fields are read at correct offsets)
- virDomainBlockCommit bandwidth parameter: long -> NativeLong
  (fixes ABI mismatch in function call)

+++ src/main/java/org/libvirt/DomainBlockJobInfo.java
@@ -13,7 +13,7 @@ public class DomainBlockJobInfo {
 
     public DomainBlockJobInfo(final virDomainBlockJobInfo virInfo) {
         type = virInfo.type;
-        bandwidth = virInfo.bandwidth;
+        bandwidth = virInfo.bandwidth.longValue();
         cur = virInfo.cur;
         end = virInfo.end;
     }
